Remove TypeNameFeature - Just use roslyn directly as we no longer have the split between language / csharp
