+++ src/Compiler/Microsoft.CodeAnalysis.Razor.Compiler/src/Language/Components/ComponentGenericTypePass.cs
@@ -3,10 +3,14 @@
 
 using System;
 using System.Collections.Generic;
+using System.Collections.Immutable;
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using Microsoft.AspNetCore.Razor.Language.Intermediate;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.Razor;
 
 namespace Microsoft.AspNetCore.Razor.Language.Components;
 
@@ -16,26 +20,9 @@ namespace Microsoft.AspNetCore.Razor.Language.Components;
 // 3. Rewrites the type names of parameters/child content to substitute generic type arguments
 internal class ComponentGenericTypePass : ComponentIntermediateNodePassBase, IRazorOptimizationPass
 {
-    private TypeNameFeature? _typeNameFeature;
-
     // Runs after components/eventhandlers/ref/bind/templates. We want to validate every component
     // and it's usage of ChildContent.
     public override int Order => 160;
-
-    private TypeNameFeature TypeNameFeature
-    {
-        get
-        {
-            // Doing lazy initialization here to avoid making things really complicated when we don't
-            // need to exercise this code in tests.
-            if (_typeNameFeature == null)
-            {
-                _typeNameFeature = GetRequiredFeature<TypeNameFeature>();
-            }
-
-            return _typeNameFeature;
-        }
-    }
 
     protected override void ExecuteCore(RazorCodeDocument codeDocument, DocumentIntermediateNode documentNode)
     {
@@ -44,21 +31,14 @@ internal class ComponentGenericTypePass : ComponentIntermediateNodePassBase, IRa
             return;
         }
 
-        var visitor = new Visitor(this);
+        var visitor = new Visitor();
         visitor.Visit(documentNode);
     }
 
-    private class Visitor : IntermediateNodeWalker
+    internal class Visitor : IntermediateNodeWalker
     {
-        private readonly ComponentGenericTypePass _pass;
-
         // Incrementing ID for type inference method names
         private int _id;
-
-        public Visitor(ComponentGenericTypePass pass)
-        {
-            _pass = pass;
-        }
 
         public override void VisitComponent(ComponentIntermediateNode node)
         {
@@ -117,7 +97,7 @@ internal class ComponentGenericTypePass : ComponentIntermediateNodePassBase, IRa
                 if (ValidateTypeArguments(node, bindings))
                 {
                     var mappings = bindings.ToDictionary(kvp => kvp.Key, kvp => kvp.Value.Content);
-                    RewriteTypeNames(_pass.TypeNameFeature.CreateGenericTypeRewriter(mappings), node, hasTypeArgumentSpecified);
+                    RewriteTypeNames(new GenericTypeNameRewriter(mappings), node, hasTypeArgumentSpecified);
                 }
 
                 return;
@@ -131,7 +111,7 @@ internal class ComponentGenericTypePass : ComponentIntermediateNodePassBase, IRa
 
             // Since we're generating code in a different namespace, we need to 'global qualify' all of the types
             // to avoid clashes with our generated code.
-            RewriteTypeNames(_pass.TypeNameFeature.CreateGlobalQualifiedTypeNameRewriter(bindings.Keys), node, hasTypeArgumentSpecified: false, bindings);
+            RewriteTypeNames(new GlobalQualifiedTypeNameRewriter(bindings.Keys), node, hasTypeArgumentSpecified: false, bindings);
 
             //
             // We need to verify that an argument was provided that 'covers' each type parameter.
@@ -145,7 +125,7 @@ internal class ComponentGenericTypePass : ComponentIntermediateNodePassBase, IRa
                     // Keep only type parameters defined by this component.
                     typeParameters = typeParameters.Where(bindings.ContainsKey).ToArray();
 
-                    var attributeValueIsLambda = _pass.TypeNameFeature.IsLambda(GetContent(attribute));
+                    var attributeValueIsLambda = SyntaxFactory.ParseExpression(GetContent(attribute)) is LambdaExpressionSyntax;
                     var provideCascadingGenericTypes = new CascadingGenericTypeParameter
                     {
                         GenericTypeNames = typeParameters,
@@ -259,7 +239,7 @@ internal class ComponentGenericTypePass : ComponentIntermediateNodePassBase, IRa
                 // helpful as possible. So let's substitute 'object' for all of those type parameters, and add
                 // an error.
                 var mappings = bindings.ToDictionary(kvp => kvp.Key, kvp => kvp.Value.Content);
-                RewriteTypeNames(_pass.TypeNameFeature.CreateGenericTypeRewriter(mappings), node, bindings: bindings);
+                RewriteTypeNames(new GenericTypeNameRewriter(mappings), node, bindings: bindings);
 
                 node.Diagnostics.Add(ComponentDiagnosticFactory.Create_GenericComponentTypeInferenceUnderspecified(node.Source, node, node.Component.GetTypeParameters()));
             }
@@ -290,7 +270,7 @@ internal class ComponentGenericTypePass : ComponentIntermediateNodePassBase, IRa
             // Two cases;
             // 1. name is a simple identifier like TItem
             // 2. name contains type parameters like Dictionary<string, TItem>
-            typeParameters = _pass.TypeNameFeature.ParseTypeParameters(globallyQualifiedTypeName ?? boundAttribute.TypeName);
+            typeParameters = ParseTypeParameters(globallyQualifiedTypeName ?? boundAttribute.TypeName);
             if (typeParameters.Count == 0)
             {
                 typeParameters = new[] { boundAttribute.TypeName };
@@ -334,8 +314,6 @@ internal class ComponentGenericTypePass : ComponentIntermediateNodePassBase, IRa
 
         private void RewriteTypeNames(TypeNameRewriter rewriter, ComponentIntermediateNode node, bool? hasTypeArgumentSpecified = null, IDictionary<string, Binding>? bindings = null)
         {
-            var typeNameFeature = _pass.TypeNameFeature;
-
             // Rewrite the component type name
             node.TypeName = rewriter.Rewrite(node.TypeName);
 
@@ -371,7 +349,8 @@ internal class ComponentGenericTypePass : ComponentIntermediateNodePassBase, IRa
                     }
                     else if(attribute.BoundAttribute?.IsEventCallbackProperty() ?? false)
                     {
-                        var typeParameters = typeNameFeature.ParseTypeParameters(attribute.TypeName);
+                        Debug.Assert(attribute.TypeName is not null);
+                        var typeParameters = ParseTypeParameters(attribute.TypeName);
                         for (int i = 0; i < typeParameters.Count; i++)
                         {
                             var parameter = typeParameters[i];
@@ -494,7 +473,68 @@ internal class ComponentGenericTypePass : ComponentIntermediateNodePassBase, IRa
 
             classNode.Children.Add(typeInferenceNode);
         }
+
+        public IReadOnlyList<string> ParseTypeParameters(string typeName)
+        {
+            if (typeName == null)
+            {
+                throw new ArgumentNullException(nameof(typeName));
+            }
+
+            var parsed = SyntaxFactory.ParseTypeName(typeName);
+
+            if (parsed is IdentifierNameSyntax identifier)
+            {
+                return Array.Empty<string>();
     }
+
+            if (TryParseCore(parsed) is { IsDefault: false } list)
+            {
+                return list;
+            }
+
+            return parsed.DescendantNodesAndSelf()
+                .OfType<TypeArgumentListSyntax>()
+                .SelectMany(arg => arg.Arguments)
+                .SelectMany(t => ParseCore(t)).ToArray();
+
+            static ImmutableArray<string> TryParseCore(TypeSyntax parsed)
+            {
+                if (parsed is ArrayTypeSyntax array)
+                {
+                    return ParseCore(array.ElementType);
+                }
+
+                if (parsed is TupleTypeSyntax tuple)
+                {
+                    return tuple.Elements.SelectManyAsArray(a => ParseCore(a.Type));
+                }
+
+                return default;
+            }
+
+            static ImmutableArray<string> ParseCore(TypeSyntax parsed)
+            {
+                // Recursively drill into arrays `T[]` and tuples `(T, T)`.
+                if (TryParseCore(parsed) is { IsDefault: false } list)
+                {
+                    return list;
+                }
+
+                // When we encounter an identifier, we assume it's a type parameter.
+                if (parsed is IdentifierNameSyntax identifier)
+                {
+                    return ImmutableArray.Create(identifier.Identifier.Text);
+                }
+
+                // Generic names like `C<T>` are ignored here because we will visit their type argument list
+                // via the `.DescendantNodesAndSelf().OfType<TypeArgumentListSyntax>()` call above.
+                return ImmutableArray<string>.Empty;
+            }
+        }
+    }
+
+
 
     private class Binding
     {
