+++ src/Compiler/Microsoft.CodeAnalysis.Razor/test/ComponentGenericTypePassParsingTest.cs
@@ -2,12 +2,13 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Linq;
+using Microsoft.AspNetCore.Razor.Language.Components;
 using Roslyn.Test.Utilities;
 using Xunit;
 
 namespace Microsoft.CodeAnalysis.Razor;
 
-public class DefaultTypeNameFeatureTest
+public class ComponentGenericTypePassParsingTest
 {
     [Theory]
     [InlineData("C", 0)]
@@ -32,7 +33,7 @@ public class DefaultTypeNameFeatureTest
     public void ParseTypeParameters(string input, int expectedNumberOfTs)
     {
         // Arrange.
-        var feature = new DefaultTypeNameFeature();
+        var feature = new ComponentGenericTypePass.Visitor();
 
         // Act.
         var parsed = feature.ParseTypeParameters(input);
