+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Formatting_NetFx/FormattingTestBase.cs
@@ -273,7 +273,6 @@ public abstract class FormattingTestBase : RazorToolingIntegrationTestBase
             builder =>
             {
                 builder.SetRootNamespace(inGlobalNamespace ? string.Empty : "Test");
-                builder.Features.Add(new DefaultTypeNameFeature());
 
                 builder.ConfigureParserOptions(builder =>
                 {
