+++ src/Razor/test/Microsoft.AspNetCore.Razor.Test.Common.Tooling/LanguageServer/LanguageServerTestBase.cs
@@ -73,8 +73,6 @@ public abstract class LanguageServerTestBase(ITestOutputHelper testOutput) : Too
             {
                 builder.UseRoslynTokenizer = true;
             });
-
-            b.Features.Add(new DefaultTypeNameFeature());
         });
 
         var importDocumentName = fileKind == RazorFileKind.Legacy ? "_ViewImports.cshtml" : "_Imports.razor";
